requirements.txt
fastapi==0.115.7
feedparser==6.0.11
loguru==0.7.2
openai==1.60.0
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.18.2
vosk==0.3.44
